add ScreenShotTaker
